docs: Create a Readme file for the project
A readme file is created in a standard form, the file contains the title, prerequisite software, a brief illustration of the code, how to run the code, and other essential information of the project.

The readme file only contains an essential information for setting up and running the code of this project. For users who want to understand the numerical scheme and the derivation and analysis of this project's model,  a separate proposal-like file (not yet finished) will be provided later.